flipping to default to unsanitized, work to subclass to follow
