+++ spec/comma/sanitized_data_extractor_spec.rb
@@ -1,3 +1,4 @@
+
 # -*- coding: utf-8 -*-
 require 'spec_helper'
 
@@ -8,13 +9,13 @@ require 'spec_helper'
 #   isbn :number_10 => 'ISBN-10', :number_13 => 'ISBN-13'
 # end
 
-describe Comma::UnsanitizedDataExtractor do
+describe Comma::SanitizedDataExtractor do
 
   before do
     @isbn = Isbn.new('123123123', '321321321')
     @book = Book.new('Smalltalk-80', 'Language and Implementation', @isbn)
 
-    @data = @book.to_comma(false)
+    @data = @book.to_comma(true)
   end
 
   describe 'when no parameters are provided' do
@@ -41,20 +42,20 @@ describe Comma::UnsanitizedDataExtractor do
       end
 
       it 'should not fail when an associated object is nil' do
-        lambda { Book.new('Smalltalk-80', 'Language and Implementation', nil).to_comma(false) }.should_not raise_error
+        lambda { Book.new('Smalltalk-80', 'Language and Implementation', nil).to_comma(true) }.should_not raise_error
       end
     end
   end
 
 end
 
-describe Comma::UnsanitizedDataExtractor, 'id attribute' do
+describe Comma::SanitizedDataExtractor, 'id attribute' do
   before do
     @data = Class.new(Struct.new(:id)) do
       comma do
         id 'ID' do |id| '42' end
       end
-    end.new(1).to_comma(false)
+    end.new(1).to_comma(true)
   end
 
   it 'id attribute should yield block' do
@@ -62,7 +63,7 @@ describe Comma::UnsanitizedDataExtractor, 'id attribute' do
   end
 end
 
-describe Comma::UnsanitizedDataExtractor, 'with static column method' do
+describe Comma::SanitizedDataExtractor, 'with static column method' do
   before do
     @data = Class.new(Struct.new(:id, :name)) do
       comma do
@@ -71,7 +72,7 @@ describe Comma::UnsanitizedDataExtractor, 'with static column method' do
         __static_column__ 'STATIC' do '' end
         __static_column__ 'STATIC' do |o| o.name end
       end
-    end.new(1, 'John Doe').to_comma(false)
+    end.new(1, 'John Doe').to_comma(true)
   end
 
   it 'should extract headers' do
@@ -79,7 +80,7 @@ describe Comma::UnsanitizedDataExtractor, 'with static column method' do
   end
 end
 
-describe Comma::UnsanitizedDataExtractor, 'nil value' do
+describe Comma::SanitizedDataExtractor, 'nil value' do
   before do
     @data = Class.new(Struct.new(:id, :name)) do
       comma do
@@ -87,7 +88,7 @@ describe Comma::UnsanitizedDataExtractor, 'nil value' do
         name 'Name'
         name 'Name' do |name| nil end
       end
-    end.new(1, nil).to_comma(false)
+    end.new(1, nil).to_comma(true)
   end
 
   it 'should extract nil' do
@@ -95,7 +96,7 @@ describe Comma::UnsanitizedDataExtractor, 'nil value' do
   end
 end
 
-describe Comma::UnsanitizedDataExtractor, 'value starting with "-", "+", "=", "@"' do
+describe Comma::SanitizedDataExtractor, 'value starting with "-", "+", "=", "@"' do
   before do
     @data = Class.new(Struct.new(:name)) do
       comma do
@@ -103,10 +104,10 @@ describe Comma::UnsanitizedDataExtractor, 'value starting with "-", "+", "=", "@
         name 'name' do |name| '-@1morestr1n6' end
         name 'name' do |name| '+1234567890' end
       end
-    end.new(1).to_comma(false)
+    end.new(1).to_comma(true)
   end
 
-  it 'not change any of the values' do
-    @data.should eq(["+somestring", "-@1morestr1n6", "+1234567890"])
+  it 'removes special characters for non digits and leaves only digits alone' do
+    @data.should eq(["somestring", "1morestr1n6", "+1234567890"])
   end
 end
